+++ upsi/ASE/plain_ASE.h
@@ -6,7 +6,7 @@ namespace upsi {
 
 class PlainASE : public ASE{
     public:
-        PlainASE(size_t _n) : ASE(_n, true) {elem_cnt = 0;}
+        PlainASE(int _n) : ASE(_n, true) {elem_cnt = 0;}
         void clear() override {elem_cnt = 0;}
         bool isEmpty() override {return elem_cnt == 0;}
         void build(const std::vector<Element>& elems, oc::block ro_seed = oc::ZeroBlock) override;
